Implemented teacher adding grades and displaying students with grades

+++ Student Management System/src/main/java/org/example/studentmanagementsystem/model/entities/Grade.java
@@ -27,7 +27,7 @@ public class Grade {
     private Subject subject;
 
     @Column(name = "grade")
-    private Double grade;
+    private Integer grade;
 
     @Column(name = "description")
     private String description;
